+++ local/features/src/common/index.ts
@@ -4,6 +4,7 @@ import { applyToastrStyles } from "@/common/toastr.js";
 import "@/common/header.js";
 import { applyCssConstants } from "@/common/css-constants.js";
 import "@/common/material-icons.js";
+import { startNotificationReadAll } from "@/common/notification-read-all.js";
 
 let started = false;
 
@@ -15,4 +16,5 @@ export function startCommon(): void {
 
   applyCssConstants();
   applyToastrStyles();
+  startNotificationReadAll();
 }